+++ .github/workflows/gotests.yml
@@ -1,0 +1,39 @@
+name: "gotests"
+
+on:
+  push:
+    branches: [ "master" ]
+  pull_request:
+    # The branches below must be a subset of the branches above
+    branches: [ "master" ]
+  schedule:
+    - cron: '29 21 * * 6'
+
+jobs:
+  oldest_supported:
+    runs-on: ubuntu-latest
+    timeout-minutes: 5
+    steps:
+      - uses: actions/checkout@v3
+      - uses: actions/setup-go@v3
+        with:
+          go-version-file: 'go.mod'
+      - name: normal tests
+        run: go test ./...
+      - name: race tests
+        run: go test -race ./...
+
+  latest:
+    runs-on: ubuntu-latest
+    timeout-minutes: 5
+    steps:
+      - uses: actions/checkout@v3
+      - uses: actions/setup-go@v3
+        with:
+          go-version: "1.x"
+          check-latest: true
+      - name: normal tests
+        run: go test ./...
+      - name: race tests
+        run: go test -race ./...
+
