+++ .travis.yml
@@ -1,8 +1,26 @@
 language: go
+
 go:
   - 1.14
   - tip
+
+notifications:
+  email: false
+
+env:
+  global:
+    # Coveralls.io token.
+    - secure: "JtfIj2ItU0o6/wxnwaCULmPnWO1fQHrX3hGbWey56hPzybA3pQEs5YOZt16nQU5GS8p3Z3OIILUP4LGebqsLU5dyWA/GxnZv45MoZ8TS8lB7aie7l5Y6yYvq0ZBJoMWxXSY6sZJ1sKJG1F7NlQoknZToxvlkfF4nP9N5RQpL8iU="
+
+install:
+  - go get -t ./...
+
 before_script:
-  - go get -d ./...
+  - wget https://github.com/mewmew/ci/raw/master/get_tools.sh
+  - chmod +x get_tools.sh
+  - ./get_tools.sh
+  - wget https://github.com/mewmew/ci/raw/master/ci_checks.sh
+  - chmod +x ci_checks.sh
+
 script:
-  - go test -v ./...
+  - ./ci_checks.sh
