+++ lib/waveform.go
@@ -1,5 +1,5 @@
 // Package waveform is capable of generating waveform images from audio streams.  MIT Licensed.
-package waveform
+package lib
 
 import (
 	"image"
