+++ lib/waveform_example_test.go
@@ -1,4 +1,4 @@
-package waveform
+package lib
 
 import (
 	"bytes"
