+++ main.go
@@ -1,0 +1,12 @@
+package main
+
+import (
+	"fmt"
+
+	"github.com/mdlayher/waveform/cmd"
+)
+
+func main() {
+	fmt.Println("Start")
+	cmd.AppRun()
+}
